Training/requirements.txt
numpy
pillow
opencv-python
aiogram
mysql
opencv_contrib_python
opencv-python-headless